Verify count of uploads before deleting

+++ app/Http/Controllers/Upload/UploadController.php
@@ -39,6 +39,11 @@ class UploadController extends Controller
     {
         $this->authorize('touch', $file);
         $this->authorize('touch', $upload);
+        
+        if ($file->uploads->count() === 1) {
+            return response()->json(null, 422);
+        }
+        
         $upload->delete();
         Storage::delete('/files/' . $file->identifier . '/' . $upload->filename);
